refactor: update random sampling methods to use u64
- Renamed getter methods in random modules to follow a consistent naming convention by prefixing with "get_".
- Changed parameter types for sampling methods from `usize` to `u64` for better compatibility with larger values.
- Updated corresponding tests to reflect these changes and ensure correctness.
- Improved code readability and maintainability by standardizing method names and types.

+++ src/utils/circulant_embedding.rs
@@ -164,8 +164,8 @@ impl CirculantEmbedding {
         };
 
         // Generate a random Gaussian vector (components for complex noise)
-        let z_real = normal::standard_rands(m);
-        let mut z_imag = normal::standard_rands(m);
+        let z_real = normal::standard_rands(m as u64);
+        let mut z_imag = normal::standard_rands(m as u64);
 
         // Special handling for real output: Z_k = conj(Z_{m-k})
         // This implies z_imag[0] = 0 and z_imag[m/2] = 0 (if m is even)
